flask-backend/requirements.txt
Flask-CORS==4.0.0
Werkzeug==3.0.1